+++ Sources/Sentry/SentryUIViewControllerPerformanceTracker.m
@@ -85,8 +85,7 @@
         callbackToOrigin:callbackToOrigin
                    block:^{
                        SENTRY_LOG_DEBUG(@"Tracking loadView");
-                       [self createTransaction:controller];
-                       [self createTimeToDisplay:controller];
+                       [self startRootSpanFor:controller];
                        [self measurePerformance:@"loadView"
                                          target:controller
                                callbackToOrigin:callbackToOrigin];
@@ -101,15 +100,14 @@
         callbackToOrigin:callbackToOrigin
                    block:^{
                        SENTRY_LOG_DEBUG(@"Tracking viewDidLoad");
-                       [self createTransaction:controller];
-                       [self createTimeToDisplay:controller];
+                       [self startRootSpanFor:controller];
                        [self measurePerformance:@"viewDidLoad"
                                          target:controller
                                callbackToOrigin:callbackToOrigin];
                    }];
 }
 
-- (void)createTransaction:(UIViewController *)controller
+- (void)startRootSpanFor:(UIViewController *)controller
 {
     SentrySpanId *spanId
         = objc_getAssociatedObject(controller, &SENTRY_UI_PERFORMANCE_TRACKER_SPAN_ID);
@@ -117,13 +115,20 @@
     // If the user manually calls loadView outside the lifecycle we don't start a new transaction
     // and override the previous id stored.
     if (spanId == nil) {
+
+        // The tracker must create a new transaction and bind it to the scope when there is no
+        // active span. If the user didn't call reportFullyDisplayed, the previous UIViewController
+        // transaction is still bound to the scope, and we need to finish and remove it from the
+        // scope so that we can bind this new UIViewController transaction to the scope.
+        if (self.tracker.activeSpanId == nil) {
+            [self.currentTTDTracker finishSpansIfNotFinished];
+        }
+
         NSString *name = [SwiftDescriptor getObjectClassName:controller];
         spanId = [self.tracker startSpanWithName:name
                                       nameSource:kSentryTransactionNameSourceComponent
                                        operation:SentrySpanOperationUILoad
                                           origin:SentryTraceOriginAutoUIViewController];
-        SENTRY_LOG_DEBUG(@"Started span with id %@ to track view controller %@.",
-            spanId.sentrySpanIdString, name);
 
         // Use the target itself to store the spanId to avoid using a global mapper.
         objc_setAssociatedObject(controller, &SENTRY_UI_PERFORMANCE_TRACKER_SPAN_ID, spanId,
@@ -133,13 +138,15 @@
         // to serve as an umbrella transaction that will capture every span
         // happening while the transaction is active.
         if (self.tracker.activeSpanId == nil) {
+            SENTRY_LOG_DEBUG(@"Started new transaction with id %@ to track view controller %@.",
+                spanId.sentrySpanIdString, name);
             [self.tracker pushActiveSpan:spanId];
+        } else {
+            SENTRY_LOG_DEBUG(@"Started child span with id %@ to track view controller %@.",
+                spanId.sentrySpanIdString, name);
         }
     }
-}
 
-- (void)createTimeToDisplay:(UIViewController *)controller
-{
     SentrySpan *vcSpan = [self viewControllerPerformanceSpan:controller];
 
     if (![vcSpan isKindOfClass:[SentryTracer self]]) {
@@ -154,6 +161,8 @@
         return;
     }
 
+    [self.currentTTDTracker finishSpansIfNotFinished];
+
     SentryTimeToDisplayTracker *ttdTracker =
         [[SentryTimeToDisplayTracker alloc] initForController:controller
                                            waitForFullDisplay:self.enableWaitForFullDisplay
@@ -162,11 +171,8 @@
     if ([ttdTracker startForTracer:(SentryTracer *)vcSpan]) {
         objc_setAssociatedObject(controller, &SENTRY_UI_PERFORMANCE_TRACKER_TTD_TRACKER, ttdTracker,
             OBJC_ASSOCIATION_ASSIGN);
-
-        [self.currentTTDTracker finishSpansIfNotFinished];
         self.currentTTDTracker = ttdTracker;
     } else {
-        [self.currentTTDTracker finishSpansIfNotFinished];
         self.currentTTDTracker = nil;
     }
 }